last changes

+++ IT2B-G05-BankApp/it2bank/src/main/java/nl/inholland/it2bank/repository/UserRepository.java
@@ -17,7 +17,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<UserModel, Long> {
 
-    default  List<UserModel> findUserByAttributes(Integer id, String firstName, String lastName, Long bsn, String phoneNumber, String email, UserRoles role, Double transactionLimit, Double dailyLimit) {
+    default  List<UserModel> findUserByAttributes(Long id, String firstName, String lastName, Long bsn, String phoneNumber, String email, UserRoles role, Double transactionLimit, Double dailyLimit) {
         return findAll(new Specification<UserModel>() {
             @Override
             public Predicate toPredicate(Root<UserModel> root, CriteriaQuery<?> query, CriteriaBuilder criteriaBuilder) {
